+++ longevity/src/main/scala/longevity/config/TestConfig.scala
@@ -4,10 +4,11 @@ package longevity.config
  *
  * @param mongodb the test configuration for MongoDB
  * @param cassandra the test configuration for Cassandra
+ * @param jdbc the JDBC configuration. used by the SQLite back end
  *
  * @see LongevityConfig
  */
 case class TestConfig(
   cassandra: CassandraConfig,
   mongodb: MongoDBConfig,
-  sqlite: SQLiteConfig)
+  jdbc: JdbcConfig)